+++ packages/terra-form/src/Control.jsx
@@ -1,47 +1,49 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
+import classNames from 'classnames/bind';
+
 import 'terra-base/lib/baseStyles';
+import styles from './Control.scss';
 
 import Input from './Input';
 
-import './Control.scss';
+const cx = classNames.bind(styles);
 
 const propTypes = {
   /**
-   * The input type
+   * The input type.
    */
   type: PropTypes.oneOf(['checkbox', 'radio']).isRequired,
   /**
-   * The checked property of the Input element. Use this to generate a controlled Control Element
+   * The checked property of the Input element. Use this to generate a controlled Control Element.
    */
   checked: PropTypes.bool,
   /**
-   * The checked property of the Input element. Use this to generate an uncontrolled Control Element
+   * The checked property of the Input element. Use this to generate an uncontrolled Control Element.
    */
   defaultChecked: PropTypes.bool,
   /**
-   * id of the input field
+   * The id of the input field.
    */
   id: PropTypes.string,
   /**
-   * Additional attributes for the input object
+   * Additional attributes for the input object.
    */
   inputAttrs: PropTypes.object,
   /**
-   * Whether the form element is inline or not
+   * Whether the form element is inline or not.
    */
   isInline: PropTypes.bool,
   /**
-   * Text of the label
+   * Text of the label.
    */
   labelText: PropTypes.node,
   /**
-   * Additional objects for the text object
+   * Additional objects for the text object.
    */
   labelTextAttrs: PropTypes.object,
   /**
-   * Name of the input attribute
+   * Name of the input attribute.
    */
   name: PropTypes.string,
   /**
@@ -49,7 +51,7 @@ const propTypes = {
    */
   onChange: PropTypes.func,
   /**
-   * The Value of the input element
+   * The value of the input element.
    */
   value: PropTypes.string,
 };
@@ -82,16 +84,21 @@ const Control = ({
   value,
   ...customProps
 }) => {
-  const labelClassNames = classNames(
-    'terra-Form-control',
-    { 'terra-Form-control--inline': isInline },
+  const labelClassNames = cx([
+    'control',
+    { 'control-inline': isInline },
     customProps.className,
-  );
+  ]);
 
-  const labelTextClasses = classNames(
-    'terra-Form-control-label-text',
+  const labelTextClasses = cx([
+    'control-label-text',
     labelTextAttrs.className,
-  );
+  ]);
+
+  const controlInputClasses = cx([
+    'input',
+    inputAttrs.className,
+  ]);
 
   const controlInputAttrs = Object.assign({}, inputAttrs);
 
@@ -110,6 +117,7 @@ const Control = ({
         type={type}
         onChange={onChange}
         {...controlInputAttrs}
+        className={controlInputClasses}
       />
       <span {...labelTextAttrs} className={labelTextClasses}>{labelText}</span>
     </label>