+++ packages/terra-form/src/Field.jsx
@@ -1,21 +1,23 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
-import 'terra-base/lib/baseStyles';
+import classNames from 'classnames/bind';
 
-import './Field.scss';
+import 'terra-base/lib/baseStyles';
+import styles from './Field.scss';
+
+const cx = classNames.bind(styles);
 
 const propTypes = {
   /**
-   * Children the Field contains
+   * Children the Field contains.
    */
   children: PropTypes.node,
   /**
-   * Error message for when the input is invalid
+   * Error message for when the input is invalid.
    */
   error: PropTypes.node,
   /**
-   * Help element to display with the input
+   * Help element to display with the input.
    */
   help: PropTypes.node,
   /**
@@ -23,19 +25,19 @@ const propTypes = {
    */
   htmlFor: PropTypes.string,
   /**
-   * Determines whether the field is an inline field
+   * Determines whether the field is an inline field.
    */
   isInline: PropTypes.bool,
   /**
-   * Label of the input
+   * Label of the input.
    */
   label: PropTypes.node,
   /**
-   * Attributes to attach to the label
+   * Attributes to attach to the label.
    */
   labelAttrs: PropTypes.object,
   /**
-   * Determines whether the field is required
+   * Determines whether the field is required.
    */
   required: PropTypes.bool,
 };
@@ -52,24 +54,24 @@ const defaultProps = {
 };
 
 const Field = ({ children, error, help, htmlFor, isInline, label, labelAttrs, required, ...customProps }) => {
-  const fieldClasses = classNames(
-    'terra-Form-field',
-    { 'terra-Form-field--inline': isInline },
-    { 'terra-Form-field--required': required },
+  const fieldClasses = cx([
+    'field',
+    { 'field-inline': isInline },
+    { 'field-required': required },
     customProps.className,
-  );
+  ]);
 
-  const labelClassNames = classNames(
-    'terra-Form-label',
+  const labelClassNames = cx([
+    'label',
     labelAttrs.className,
-  );
+  ]);
 
   return (
     <div {...customProps} className={fieldClasses}>
       {label && <label htmlFor={htmlFor} {...labelAttrs} className={labelClassNames}>{label}</label>}
       {children}
-      {help && <small className="terra-Form-helpText" tabIndex="-1">{help}</small>}
-      {error && <small className="terra-Form-error" tabIndex="-1">{error}</small>}
+      {help && <small className={cx('help-text')} tabIndex="-1">{help}</small>}
+      {error && <small className={cx('error')} tabIndex="-1">{error}</small>}
     </div>
   );
 };