+++ packages/terra-form/src/Textarea.jsx
@@ -1,9 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
-import 'terra-base/lib/baseStyles';
+import classNames from 'classnames/bind';
 
-import './Textarea.scss';
+import 'terra-base/lib/baseStyles';
+import styles from './Textarea.scss';
+
+const cx = classNames.bind(styles);
 
 const propTypes = {
   /**
@@ -11,7 +13,7 @@ const propTypes = {
    */
   defaultValue: PropTypes.string,
   /**
-   * Name of the input
+   * Name of the input.
    */
   name: PropTypes.string,
   /**
@@ -19,7 +21,7 @@ const propTypes = {
    */
   onChange: PropTypes.func,
   /**
-   * Whether the input is required or not
+   * Whether the input is required or not.
    */
   required: PropTypes.bool,
   /**
@@ -45,7 +47,7 @@ const Textarea = ({
   ...customProps
 }) => {
   const additionalTextareaProps = Object.assign({}, customProps);
-  const textareaClasses = classNames('terra-Form-textarea', additionalTextareaProps.className);
+  const textareaClasses = cx(['textarea', additionalTextareaProps.className]);
 
   if (required) {
     additionalTextareaProps['aria-required'] = 'true';