+++ packages/terra-form/tests/nightwatch/input/input-spec.js
@@ -17,21 +17,9 @@ module.exports = {
       .assert.elementPresent('input');
   },
 
-  'Properly initializes Input as uncontrolled when passed defaultValue': (browser) => {
+  'Properly initializes Input with the correct attributes': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/form-tests/input/populated_uncontrolled`)
-      .assert.elementPresent('input[name="foo"][value="bar"][required]')
-      .clearValue('input[name="foo"]')
-      .setValue('input[name="foo"][value="bar"]', 'New Value')
-      .assert.value('input[name="foo"]', 'New Value');
-  },
-
-  'Properly initializes Input as controlled when passed value and onChange': (browser) => {
-    browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/form-tests/input/populated_controlled`)
-      .assert.elementPresent('input[name="foo"][value="Jack"][required]')
-      .clearValue('input[name="foo"]')
-      .setValue('input[name="foo"]', 'New Value')
-      .assert.value('input[name="foo"]', 'New Value');
+      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/form-tests/input/populated`)
+      .assert.elementPresent('input[name="foo"][value="bar"][required]');
   },
 };