+++ packages/terra-form/tests/nightwatch/select-field/DefaultSelectField.jsx
@@ -2,6 +2,6 @@ import React from 'react';
 
 import SelectField from '../../../lib/SelectField';
 
-const selectfield = () => (<SelectField choices={['moo']} />);
+const selectfield = () => (<SelectField id="default-select-field" choices={['moo']} />);
 
 export default selectfield;