+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added requestFocus and releaseFocus props to support requesting and releasing focus.
+
+### Changed
+* Converted component to use CSS modules
 
 1.2.0 - (July 26, 2017)
 ------------------