+++ packages/terra-popup/CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added requestFocus and releaseFocus props to support requesting and releasing focus.
+
+### Fixed
+* Prevent Invalid Props from breaking size logic. (Issue #679)
 
 1.0.0 - (July 26, 2017)
 ------------------